requirements.txt
requests
beautifulsoup4
PyExecJs